edited base scene and subtitles
