+++ package.json
@@ -17,6 +17,7 @@
   },
   "homepage": "https://github.com/cmr624/cm-phaser-library#readme",
   "dependencies": {
-    "phaser": "^3.22.0"
+    "phaser": "^3.22.0",
+    "phaser3-rex-plugins": "^1.1.13"
   }
 }
