+++ src/edu/baseEduScene.ts
@@ -1,8 +1,8 @@
-import SubtitleManager from "./subtitleManager";
-import { BaseContainer } from "../base/container";
+import SubtitleManager, {SubtitleContainer} from "./subtitleManager";
 
 abstract class EDUSceneTemplate extends Phaser.Scene {
     subtitleManager! : SubtitleManager;
+    rexUI!:any;
 }
 
 export default class BaseEDUScene extends EDUSceneTemplate {
@@ -18,6 +18,7 @@ export default class BaseEDUScene extends EDUSceneTemplate {
     }
 
     initializeSubtitleManager(){
+        //should load from JSON
         let dict : Map<string, string> = new Map<string, string>();
         dict['correct'] = '(Correct Sound)';
         dict['goodWork'] = 'Good Work!';
@@ -28,32 +29,6 @@ export default class BaseEDUScene extends EDUSceneTemplate {
         this.initializeSubtitleManager();
         this.subtitleContainer = new SubtitleContainer(this, 800, 1000).setDepth(99);
         //this.add.container(800, 1000).setDepth(99);//.setScrollFactor(0);       
-
     }
 }
 
-export class SubtitleContainer extends BaseContainer {
-    stack : Array<Phaser.GameObjects.Text>;
-
-    constructor(scene, x, y, children?){
-        super(scene, x, y, children!);
-        this.stack = new Array<Phaser.GameObjects.Text>();
-    }
-
-    addSubtitle(text : Phaser.GameObjects.Text){
-        if (this.stack.length > 0) {
-            //move everyone on the stack up by 50 px
-            this.stack.forEach((e) => e.y -= 50);
-        }
-        this.stack.push(text);
-        this.add(text);
-        setTimeout(() => {
-            this.stack.reverse();
-            this.stack.pop()?.destroy();
-            this.stack.reverse();
-        }, 2000);
-
-    }
-
-
-}
